sap_scrapper2/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
pandas>=2.1.1
lxml>=4.9.3
urllib3>=2.0.7
python-dateutil>=2.8.2
pytz>=2023.3
litellm>=1.0.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
asyncio>=3.4.3 